+++ include/Common.h
@@ -14,6 +14,7 @@ extern TSentry Sentry;
 #include <mutex>
 #include <sstream>
 #include <zlib.h>
+#include <filesystem>
 
 #include "Compat.h"
 
@@ -45,7 +46,7 @@ public:
         std::string SSLKeyPath { "./.ssl/HttpServer/key.pem" };
         std::string SSLCertPath { "./.ssl/HttpServer/cert.pem" };
         bool HTTPServerEnabled { false };
-        int MaxPlayers { 10 };
+        int MaxPlayers { 8 };
         bool Private { true };
         int MaxCars { 1 };
         bool DebugModeEnabled { false };
@@ -56,6 +57,7 @@ public:
         bool SendErrorsMessageEnabled { true };
         int HTTPServerPort { 8080 };
         bool HTTPServerUseSSL { true };
+        bool HideUpdateMessages { false };
         [[nodiscard]] bool HasCustomIP() const { return !CustomIP.empty(); }
     };