+++ Changelog.md
@@ -6,10 +6,19 @@
 - ADDED FS.ListFiles and FS.ListDirectories
 - FIXED issue with client->server events which contain ':'
 
+# v3.0.2
+
+- ADDED Periodic update message if a new server is released
+- CHANGED Default MaxPlayers to 8
+- FIXED `MP.CreateEventTimer` filling up the queue (see https://wiki.beammp.com/en/Scripting/new-lua-scripting#mpcreateeventtimerevent_name-string-interval_ms-number-strategy-number-since-v302)
+- FIXED `MP.TriggerClientEvent` not kicking the client if it failed
+- FIXED Lua result queue handling not checking all results
+
 # v3.0.1
 
 - ADDED Backup URLs to UpdateCheck (will fail less often now)
 - ADDED console cursor left and right movement (with arrow keys) and working HOME and END key (via github.com/lionkor/commandline)
+- FIXED infinite snowmen / infinite unicycle spawning bug
 - FIXED a bug where, when run with --working-directory, the Server.log would still be in the original directory
 - FIXED a bug which could cause the plugin reload thread to spin at 100% if the reloaded plugin's didn't terminate
 - FIXED an issue which would cause servers to crash on mod download via SIGPIPE on POSIX